Finishing up everything

+++ scripts/init-db.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS bronze_data.weather_history_forecast (
     snapshot_time TIMESTAMP NOT NULL,
     latitude DECIMAL(9, 6) NOT NULL,
     longitude DECIMAL(9, 6) NOT NULL,
-    temperature DECIMAL(5, 2) NOT NULL,
-    wind_speed DECIMAL(5, 2) NOT NULL,
+    temperature DOUBLE PRECISION NOT NULL,
+    wind_speed DOUBLE PRECISION NOT NULL
 );
